chore: move tests to ts-jest
