--- .babelrc
@@ -1,6 +1,0 @@
-{
-    "plugins": [
-        "transform-async-to-generator",
-        "transform-es2015-modules-commonjs",
-    ]
-}
