+++ package.json
@@ -4,8 +4,8 @@
   "description": "encrypted sequelize fields",
   "main": "index.js",
   "scripts": {
-    "format": "npx prettier --write '*.js' 'test/*.js'",
-    "test": "mocha"
+    "format": "npx prettier --write '*.js' 'test/*.?s'",
+    "test": "jest --detectOpenHandles --forceExit"
   },
   "repository": {
     "type": "git",
@@ -22,13 +22,11 @@
   },
   "homepage": "https://github.com/defunctzombie/sequelize-encrypted",
   "devDependencies": {
-    "babel-plugin-transform-async-to-generator": "6.8.0",
-    "babel-plugin-transform-es2015-modules-commonjs": "6.10.3",
-    "babel-polyfill": "6.9.1",
-    "babel-preset-es2015": "6.9.0",
-    "babel-register": "6.9.0",
-    "mocha": "2.0.1",
+    "@types/jest": "^24.0.18",
+    "@types/sequelize": "^4.28.5",
+    "jest": "^24.9.0",
     "pg": "6.0.1",
-    "sequelize": "3.23.4"
+    "sequelize": "3.23.4",
+    "ts-jest": "^24.1.0"
   }
 }
