--- test/mocha.opts
@@ -1,5 +1,0 @@
---reporter spec
---check-leaks
---bail
---compilers js:babel-register
---require babel-polyfill
